requirements.txt
streamlit>=1.28.0
python-chess>=1.10.0
stockfish>=3.28.0

